Reworked single config file into multiple config files Amended imports according to new config files and paths
