+++ src/configs/layerGroups.ts
@@ -1,0 +1,164 @@
+import { LayerGroup } from "../interfaces"
+
+export const layerGroups:LayerGroup[] = [
+  {
+    name: "Other",
+    layers: [
+      {
+        name: "LS - Geneva/Zurich Boundary",
+        id: "LSASBDRY",
+        data: "",
+      },
+      {
+        name: "LI - Italy ARO Boundary",
+        id: "LIMMBDRY",
+        data: "",
+      },
+      {
+        name: "LS - Drone Areas",
+        id: "LSASDRONE",
+        data: "",
+      }
+    ]
+  },
+  {
+    name: "VFR Reporting Points",
+    layers: [
+      {
+        name: "LD - Croatia",
+        id: "LDZOREP",
+        data: "",
+      },
+      {
+        name: "LJ - Slovenia",
+        id: "LJLAREP",
+        data: "",
+      },
+    ]
+  },
+  {
+    name: "CTR",
+    layers: [
+      {
+        name: "EB - Belgium & Luxembourg",
+        id: "EBBUCTR",
+        data: "",
+      },
+      {
+        name: "EG - United Kingdom",
+        id: "EGXXCTR",
+        data: "",
+      },
+      {
+        name: "LD - Croatia",
+        id: "LDZOCTR",
+        data: "",
+      },
+      {
+        name: "LK - Czech Republic",
+        id: "LKAACTR",
+        data: "",
+      },
+    ]
+  },
+  {
+    name: "TMA",
+    layers: [
+      {
+        name: "EB - Belgium & Luxembourg",
+        id: "EBBUTMA",
+        data: "",
+      },
+      {
+        name: "LD - Croatia",
+        id: "LDZOTMA",
+        data: "",
+      },
+      {
+        name: "LK - Czech Republic",
+        id: "LKAATMA",
+        data: "",
+      },
+      {
+        name: "LY - Serbia and Montenegro",
+        id: "LYPRTMA",
+        data: "",
+      },
+    ]
+  },
+  {
+    name: "FIR",
+    layers: [
+      {
+        name: "BI - Iceland",
+        id: "BIRDFIR",
+        data: "",
+      },
+      {
+        name: "EB - Belgium & Luxembourg",
+        id: "EBBUFIR",
+        data: "",
+      },
+      {
+        name: "LK - Czech Republic",
+        id: "LKAAFIR",
+        data: "",
+      },
+      {
+        name: "ED - Germany",
+        id: "EDXXFIR",
+        data: "",
+      },
+      {
+        name: "EG - United Kingdom",
+        id: "EGXXFIR",
+        data: "",
+      },
+      {
+        name: "EH - Netherlands",
+        id: "EHAAFIR",
+        data: "",
+      },
+      {
+        name: "EI - Ireland",
+        id: "EISNFIR",
+        data: "",
+      },
+      {
+        name: "LA - Albania",
+        id: "LAAAFIR",
+        data: "",
+      },
+      {
+        name: "LD - Croatia",
+        id: "LDZOFIR",
+        data: "",
+      },
+      {
+        name: "LE - Spain",
+        id: "LEXXFIR",
+        data: "",
+      },
+      {
+        name: "LF - France",
+        id: "LFXXFIR",
+        data: "",
+      },
+      {
+        name: "LI - Italy",
+        id: "LIXXFIR",
+        data: "",
+      },
+      {
+        name: "LS - Switzerland",
+        id: "LSASFIR",
+        data: "",
+      },
+      {
+        name: "LY - Serbia and Montenegro",
+        id: "LYPRFIR",
+        data: "",
+      },
+    ]
+  },
+]
